map support

+++ app/src/main/java/com/fshou/ceritain/data/remote/response/StoriesResponse.kt
@@ -32,11 +32,11 @@ data class Story(
 	val description: String? = null,
 
 	@field:SerializedName("lon")
-	val lon: Float? = null,
+	val lon: Double? = null,
 
 	@field:SerializedName("id")
 	val id: String? = null,
 
 	@field:SerializedName("lat")
-	val lat: Float? = null
+	val lat: Double? = null
 ) : Parcelable
